chore: rename catalog module
